[ivp_utility] x64 ready memory allocators
- Use size_t instead of int for memory sizes
- Use uintp instead of intp for address arithmetic
- Add null check in ivp_malloc_aligned
- Add ivp_calloc_aligned implementation
- Add U suffix to memory alignment constants
- Fix magic_number type to unsigned int

Ported from source-physics commit f500517

+++ ivp_utility/ivu_types.hxx
@@ -325,13 +325,13 @@ enum IVP_RETURN_TYPE
 
 #include <string.h>
 
-char *IVP_CDECL p_calloc(int nelem, int size);
-void *IVP_CDECL p_realloc(void *memblock, int size);
-void *IVP_CDECL p_malloc(unsigned int size);
+char *IVP_CDECL p_calloc(size_t nelem, size_t size);
+void *IVP_CDECL p_realloc(void *memblock, size_t size);
+void *IVP_CDECL p_malloc(size_t size);
 void IVP_CDECL p_free(void *data);
 
-void *IVP_CDECL ivp_malloc_aligned(int size, int alignment);
-void *IVP_CDECL ivp_calloc_aligned(int size, int alignment);
+void *IVP_CDECL ivp_malloc_aligned(size_t size, unsigned short alignment);
+void *IVP_CDECL ivp_calloc_aligned(size_t size, unsigned short alignment);
 
 void IVP_CDECL ivp_free_aligned(void *data);
